requirements.txt
pandas
requests
duckdb
pyyaml
pytest